added filedata class, readfile class and words.txt
The readfile class is set up to allow the user to read information from text documents and store them to an array. The filedata class contains all of the function calls to use readfile as well as most of the backend so far. Included is also a text document containing various words.